'edit_timeline' view moved to timeline module
